+++ CHANGES.md
@@ -70,9 +70,9 @@
   - @zztkm
 - [UPDATE] CA 証明書を指定できるようにする
   - この証明書は以下のタイミングで利用される
-    - シグナリング接続時
+    - WebSocket シグナリング利用時
     - TURN-TLS 利用時
-  - `SoraMediaChannel` に `caCertificate: Certificate?` を追加する
+  - `SoraMediaChannel` に `caCertificate: X509Certificate?` を追加する
   - `SoraMediaChannel` で CA 証明書を指定しない場合は、サーバー証明書の検証にシステムのデフォルトが利用される
   - @zztkm
 - [ADD] サイマルキャストの映像のエンコーディングパラメーター `scaleResolutionDownTo` を追加する